requirements.txt
meshgpt_pytorch==0.6.7
pytorch-custom-utils==0.0.21
accelerate>=0.25.0
beartype
classifier-free-guidance-pytorch==0.5.1
einops>=0.7.0
ema-pytorch
pytorch-warmup
torch_geometric
torchtyping
vector-quantize-pytorch==1.12.8
x-transformers==1.26.6
tqdm
matplotlib
wandb
pyrr
trimesh
opencv-python
pyrender
open3d-python
easydict
chardet
deepspeed
omegaconf
scikit-image
setuptools
pytorch_lightning
mesh2sdf
numpy==1.26.4
point-cloud-utils